[JBRes-10122] Extract shared mcp-steroid e2e window helpers
Move the mcp-steroid GUI helpers (window listing, per-window screenshots,
modal-dialog dismissal, the IdeWindow model) out of test_pycharm_mcp_steroid.py
into e2e-tests/utils/mcp_steroid.py so the upcoming IDEA virtual-display test can
reuse them. Behavior-preserving refactor of the PyCharm test (list_windows now
returns an empty result on a non-zero exit, matching the old exit-code guard).
